requirements.txt
pytest
pylint
python-dotenv
boto3
psycopg2-binary
confluent-kafka